Adding local freezing link when creating request_send_funds in control.
